+++ matjalalzz/src/main/java/shop/matjalalzz/domain/user/dao/UserRepository.java
@@ -1,12 +1,13 @@
-package shop.matjalalzz.user.dao;
+package shop.matjalalzz.domain.user.dao;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shop.matjalalzz.user.domain.User;
+import shop.matjalalzz.domain.user.domain.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
+    Optional<User> findByOauthId(String oauthId);
     Optional<User> findByEmail(String email);
 }
