+++ matjalalzz/src/main/java/shop/matjalalzz/domain/user/domain/enums/Gender.java
@@ -1,0 +1,5 @@
+package shop.matjalalzz.domain.user.domain.enums;
+
+public enum Gender {
+    M, W
+}
