--- matjalalzz/src/main/java/shop/matjalalzz/user/domain/enums/Gender.java
@@ -1,5 +1,0 @@
-package shop.matjalalzz.user.domain.enums;
-
-public enum Gender {
-    M, W
-}
