--- matjalalzz/src/main/java/shop/matjalalzz/user/domain/enums/Role.java
@@ -1,6 +1,0 @@
-package shop.matjalalzz.user.domain.enums;
-
-public enum Role {
-    ADMIN, USER, OWNER
-
-}
